requirements.txt
aiohttp==3.8.4
django-celery-beat==2.2.1
django-cors-headers==3.12.0
django-model-utils==4.2.0
django-redis==5.2.0
djangorestframework-simplejwt==5.0.0
flake8==5.0.4
gunicorn==20.1.0
pip-chill==1.0.1
pre-commit==2.20.0
psycopg2-binary==2.9.5
python-decouple==3.8
urllib3==1.26.15
xmltodict==0.13.0
# aiosignal==1.3.1 # Installed as dependency for aiohttp
# amqp==5.1.1 # Installed as dependency for kombu
# asgiref==3.6.0 # Installed as dependency for django
# async-timeout==4.0.2 # Installed as dependency for aiohttp
# attrs==22.2.0 # Installed as dependency for aiohttp
# billiard==3.6.4.0 # Installed as dependency for celery
# celery==5.2.3 # Installed as dependency for django-celery-beat
# cfgv==3.3.1 # Installed as dependency for pre-commit
# charset-normalizer==3.1.0 # Installed as dependency for aiohttp
# click==8.1.3 # Installed as dependency for click-didyoumean, click-plugins, click-repl, celery
# click-didyoumean==0.3.0 # Installed as dependency for celery
# click-plugins==1.1.1 # Installed as dependency for celery
# click-repl==0.2.0 # Installed as dependency for celery
# colorama==0.4.6 # Installed as dependency for click
# deprecated==1.2.13 # Installed as dependency for redis
# distlib==0.3.6 # Installed as dependency for virtualenv
# django==3.2.18 # Installed as dependency for django-model-utils, djangorestframework-simplejwt, django-timezone-field, django-cors-headers, django-redis, djangorestframework, django-celery-beat
# django-timezone-field==4.2.3 # Installed as dependency for django-celery-beat
# djangorestframework==3.13.1 # Installed as dependency for djangorestframework-simplejwt
# filelock==3.10.6 # Installed as dependency for virtualenv
# frozenlist==1.3.3 # Installed as dependency for aiosignal, aiohttp
# identify==2.5.22 # Installed as dependency for pre-commit
# idna==3.4 # Installed as dependency for yarl
# kombu==5.2.4 # Installed as dependency for celery
# mccabe==0.7.0 # Installed as dependency for flake8
# multidict==6.0.4 # Installed as dependency for yarl, aiohttp
# nodeenv==1.7.0 # Installed as dependency for pre-commit
# packaging==23.0 # Installed as dependency for redis
# platformdirs==3.2.0 # Installed as dependency for virtualenv
# prompt-toolkit==3.0.38 # Installed as dependency for click-repl
# pycodestyle==2.9.1 # Installed as dependency for flake8
# pyflakes==2.5.0 # Installed as dependency for flake8
# pyjwt==2.6.0 # Installed as dependency for djangorestframework-simplejwt
# python-crontab==2.7.1 # Installed as dependency for django-celery-beat
# python-dateutil==2.8.2 # Installed as dependency for python-crontab
# pytz==2023.2 # Installed as dependency for django, django-timezone-field, celery, djangorestframework
# pyyaml==6.0 # Installed as dependency for pre-commit
# redis==4.1.4 # Installed as dependency for django-redis
# six==1.16.0 # Installed as dependency for click-repl, python-dateutil
# sqlparse==0.4.3 # Installed as dependency for django
# toml==0.10.2 # Installed as dependency for pre-commit
# vine==5.0.0 # Installed as dependency for kombu, amqp, celery
# virtualenv==20.21.0 # Installed as dependency for pre-commit
# wcwidth==0.2.6 # Installed as dependency for prompt-toolkit
# wrapt==1.15.0 # Installed as dependency for deprecated
# yarl==1.8.2 # Installed as dependency for aiohttp
